refactor: Move resolve logic to a filter pipeline
