+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-governance/resource-transform/istio-resource-transform/src/main/java/com/alibaba/cloud/governance/istio/protocol/impl/RdsProtocol.java
@@ -17,33 +17,15 @@
 package com.alibaba.cloud.governance.istio.protocol.impl;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import com.alibaba.cloud.commons.governance.event.LabelRoutingDataChangedEvent;
-import com.alibaba.cloud.commons.governance.labelrouting.LabelRouteRule;
-import com.alibaba.cloud.commons.governance.labelrouting.MatchService;
-import com.alibaba.cloud.commons.governance.labelrouting.UnifiedRouteDataStructure;
-import com.alibaba.cloud.commons.governance.labelrouting.rule.HeaderRule;
-import com.alibaba.cloud.commons.governance.labelrouting.rule.RouteRule;
-import com.alibaba.cloud.commons.governance.labelrouting.rule.UrlRule;
-import com.alibaba.cloud.commons.lang.StringUtils;
-import com.alibaba.cloud.commons.matcher.StringMatcher;
-import com.alibaba.cloud.commons.matcher.StringMatcherType;
 import com.alibaba.cloud.governance.istio.XdsChannel;
 import com.alibaba.cloud.governance.istio.XdsConfigProperties;
 import com.alibaba.cloud.governance.istio.XdsScheduledThreadPool;
 import com.alibaba.cloud.governance.istio.constant.IstioConstants;
+import com.alibaba.cloud.governance.istio.filter.XdsResolveFilter;
 import com.alibaba.cloud.governance.istio.protocol.AbstractXdsProtocol;
-import com.alibaba.cloud.governance.istio.util.ConvUtil;
-import io.envoyproxy.envoy.config.route.v3.HeaderMatcher;
-import io.envoyproxy.envoy.config.route.v3.QueryParameterMatcher;
-import io.envoyproxy.envoy.config.route.v3.Route;
 import io.envoyproxy.envoy.config.route.v3.RouteConfiguration;
-import io.envoyproxy.envoy.config.route.v3.RouteMatch;
-import io.envoyproxy.envoy.config.route.v3.VirtualHost;
-import io.envoyproxy.envoy.config.route.v3.WeightedCluster;
 import io.envoyproxy.envoy.service.discovery.v3.DiscoveryResponse;
 
 /**
@@ -52,21 +34,16 @@ import io.envoyproxy.envoy.service.discovery.v3.DiscoveryResponse;
  */
 public class RdsProtocol extends AbstractXdsProtocol<RouteConfiguration> {
 
-	/**
-	 * useless rds.
-	 */
-	private static final String ALLOW_ANY = "allow_any";
-
-	private static final String HEADER = "header";
-
-	private static final String PARAMETER = "parameter";
-
-	private static final String PATH = "path";
-
 	public RdsProtocol(XdsChannel xdsChannel,
 			XdsScheduledThreadPool xdsScheduledThreadPool,
-			XdsConfigProperties xdsConfigProperties) {
+			XdsConfigProperties xdsConfigProperties,
+			List<XdsResolveFilter<List<RouteConfiguration>>> rdsFilters) {
 		super(xdsChannel, xdsScheduledThreadPool, xdsConfigProperties);
+		for (XdsResolveFilter<List<RouteConfiguration>> filter : rdsFilters) {
+			if (IstioConstants.RDS_URL.equals(filter.getTypeUrl())) {
+				filters.add(filter);
+			}
+		}
 	}
 
 	@Override
@@ -81,153 +58,8 @@ public class RdsProtocol extends AbstractXdsProtocol<RouteConfiguration> {
 				log.error("unpack cluster failed", e);
 			}
 		}
+		fireXdsFilters(routes);
 		return routes;
-	}
-
-	public void resolveLabelRouting(List<RouteConfiguration> routeConfigurations) {
-		if (routeConfigurations == null) {
-			return;
-		}
-		Map<String, UnifiedRouteDataStructure> untiedRouteDataStructures = new HashMap<>();
-		for (RouteConfiguration routeConfiguration : routeConfigurations) {
-			List<VirtualHost> virtualHosts = routeConfiguration.getVirtualHostsList();
-			for (VirtualHost virtualHost : virtualHosts) {
-				UnifiedRouteDataStructure unifiedRouteDataStructure = new UnifiedRouteDataStructure();
-				String targetService = "";
-				String[] serviceAndPort = virtualHost.getName().split(":");
-				if (serviceAndPort.length > 0) {
-					targetService = serviceAndPort[0].split("\\.")[0];
-				}
-				if (ALLOW_ANY.equals(targetService)) {
-					continue;
-				}
-				unifiedRouteDataStructure.setTargetService(targetService);
-				List<Route> routes = virtualHost.getRoutesList();
-				LabelRouteRule labelRouteRule = getLabelRouteData(routes);
-				unifiedRouteDataStructure.setLabelRouteRule(labelRouteRule);
-				untiedRouteDataStructures.put(unifiedRouteDataStructure.getTargetService(),
-						unifiedRouteDataStructure);
-			}
-		}
-		applicationContext.publishEvent(
-				new LabelRoutingDataChangedEvent(untiedRouteDataStructures.values()));
-	}
-
-	private LabelRouteRule getLabelRouteData(List<Route> routes) {
-		List<MatchService> matchServices = new ArrayList<>();
-		LabelRouteRule labelRouteRule = new LabelRouteRule();
-		for (Route route : routes) {
-			String cluster = route.getRoute().getCluster();
-			if (StringUtils.isNotEmpty(cluster)) {
-				MatchService matchService = getMatchService(route, cluster, 100);
-				matchServices.add(matchService);
-			}
-			WeightedCluster weightedCluster = route.getRoute().getWeightedClusters();
-			for (WeightedCluster.ClusterWeight clusterWeight : weightedCluster
-					.getClustersList()) {
-				MatchService matchService = getMatchService(route,
-						clusterWeight.getName(), clusterWeight.getWeight().getValue());
-				matchServices.add(matchService);
-			}
-		}
-		labelRouteRule.setMatchRouteList(matchServices);
-		if (!matchServices.isEmpty()) {
-			labelRouteRule.setDefaultRouteVersion(
-					matchServices.get(matchServices.size() - 1).getVersion());
-		}
-		return labelRouteRule;
-	}
-
-	private MatchService getMatchService(Route route, String cluster, int weight) {
-		String version = "";
-		try {
-			String[] info = cluster.split("\\|");
-			version = info[2];
-		}
-		catch (Exception e) {
-			log.error("invalid cluster info for route {}", route.getName());
-		}
-		MatchService matchService = new MatchService();
-		matchService.setVersion(version);
-		matchService.setRuleList(match2RouteRules(route.getMatch()));
-		matchService.setWeight(weight);
-		return matchService;
-	}
-
-	private List<RouteRule> match2RouteRules(RouteMatch routeMatch) {
-		List<RouteRule> routeRules = new ArrayList<>();
-		for (HeaderMatcher headerMatcher : routeMatch.getHeadersList()) {
-			HeaderRule headerRule = headerMatcher2HeaderRule(headerMatcher);
-			if (headerRule != null) {
-				routeRules.add(headerRule);
-			}
-		}
-
-		for (QueryParameterMatcher parameterMatcher : routeMatch
-				.getQueryParametersList()) {
-			UrlRule.Parameter parameter = parameterMatcher2ParameterRule(
-					parameterMatcher);
-			if (parameter != null) {
-				routeRules.add(parameter);
-			}
-		}
-
-		UrlRule.Path path = new UrlRule.Path();
-		path.setType(PATH);
-		switch (routeMatch.getPathSpecifierCase()) {
-		case PREFIX:
-			path.setCondition(StringMatcherType.PREFIX.toString());
-			path.setValue(routeMatch.getPrefix());
-			break;
-
-		case PATH:
-			path.setCondition(StringMatcherType.EXACT.toString());
-			path.setValue(routeMatch.getPath());
-			break;
-
-		case SAFE_REGEX:
-			path.setCondition(StringMatcherType.REGEX.toString());
-			path.setValue(routeMatch.getSafeRegex().getRegex());
-			break;
-
-		default:
-			// unknown type
-			path = null;
-
-		}
-		if (path != null) {
-			routeRules.add(path);
-		}
-		return routeRules;
-	}
-
-	private UrlRule.Parameter parameterMatcher2ParameterRule(
-			QueryParameterMatcher queryParameterMatcher) {
-		UrlRule.Parameter parameter = new UrlRule.Parameter();
-		StringMatcher stringMatcher = ConvUtil
-				.convStringMatcher(queryParameterMatcher.getStringMatch());
-		if (stringMatcher != null) {
-			parameter.setCondition(stringMatcher.getType().toString());
-			parameter.setKey(queryParameterMatcher.getName());
-			parameter.setValue(stringMatcher.getMatcher());
-			parameter.setType(PARAMETER);
-			return parameter;
-		}
-		return null;
-	}
-
-	private HeaderRule headerMatcher2HeaderRule(HeaderMatcher headerMatcher) {
-		StringMatcher stringMatcher = ConvUtil
-				.convStringMatcher(ConvUtil.headerMatch2StringMatch(headerMatcher));
-		if (stringMatcher != null) {
-			HeaderRule headerRule = new HeaderRule();
-			headerRule.setCondition(stringMatcher.getType().toString());
-			headerRule.setKey(headerMatcher.getName());
-			headerRule.setValue(stringMatcher.getMatcher());
-			headerRule.setType(HEADER);
-			return headerRule;
-		}
-		return null;
 	}
 
 	@Override
